backend/requirements.txt
# ReimagineAI Backend Dependencies
# FastAPI & Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# AI Integration
openai>=1.12.0
google-genai>=1.0.0  # Gemini API for image editing

# Database (optional for MVP)
sqlalchemy>=2.0.25
# asyncpg>=0.29.0  # Uncomment when using PostgreSQL
alembic>=1.13.1

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.3
pydantic-settings>=2.1.0

# File handling
aiofiles>=23.2.1
pillow>=10.0.0

# HTTP Client
httpx>=0.26.0

# Environment variables
python-dotenv>=1.0.0

# Redis (for caching - optional for now)
# redis>=5.0.1
